+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialActorChannel.h
@@ -156,8 +156,13 @@ public:
 	}
 
 	// Indicates whether this client worker has "ownership" (authority over Client endpoint) over the entity corresponding to this channel.
-	FORCEINLINE bool IsOwnedByWorker() const
+	FORCEINLINE bool IsAuthoritativeClient() const
 	{
+		if (GetDefault<USpatialGDKSettings>()->bEnableResultTypes)
+		{
+			return NetDriver->StaticComponentView->HasAuthority(EntityId, SpatialConstants::GetClientAuthorityComponent(GetDefault<USpatialGDKSettings>()->bUseRPCRingBuffers));
+		}
+
 		const TArray<FString>& WorkerAttributes = NetDriver->Connection->GetWorkerAttributes();
 
 		if (const SpatialGDK::EntityAcl* EntityACL = NetDriver->StaticComponentView->GetComponentData<SpatialGDK::EntityAcl>(EntityId))
@@ -180,7 +185,7 @@ public:
 		return false;
 	}
 
-	FORCEINLINE bool IsAuthoritativeServer()
+	FORCEINLINE bool IsAuthoritativeServer() const
 	{
 		return NetDriver->IsServer() && NetDriver->StaticComponentView->HasAuthority(EntityId, SpatialConstants::POSITION_COMPONENT_ID);
 	}
@@ -206,7 +211,7 @@ public:
 	virtual int64 Close(EChannelCloseReason Reason) override;
 	// End UChannel interface
 
-	// Begin UActorChannel inteface
+	// Begin UActorChannel interface
 	virtual int64 ReplicateActor() override;
 #if ENGINE_MINOR_VERSION <= 22
 	virtual void SetChannelActor(AActor* InActor) override;
@@ -267,6 +272,7 @@ private:
 	FHandoverChangeState GetHandoverChangeList(TArray<uint8>& ShadowData, UObject* Object);
 	
 	void UpdateEntityACLToNewOwner();
+	void UpdateInterestBucketComponentId();
 
 public:
 	// If this actor channel is responsible for creating a new entity, this will be set to true once the entity creation request is issued.