+++ SpatialGDK/Source/SpatialGDK/Public/SpatialGDKSettings.h
@@ -42,6 +42,18 @@ namespace EServicesRegion
 	};
 }
 
+USTRUCT(BlueprintType)
+struct FDistanceFrequencyPair
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, EditAnywhere, Category = "SpatialGDK")
+	float DistanceRatio;
+
+	UPROPERTY(BlueprintReadOnly, EditAnywhere, Category = "SpatialGDK")
+	float Frequency;
+};
+
 UCLASS(config = SpatialGDKSettings, defaultconfig)
 class SPATIALGDK_API USpatialGDKSettings : public UObject
 {
@@ -178,11 +190,11 @@ public:
 	UPROPERTY(config)
 	bool bEnableServerQBI;
 
-	/** EXPERIMENTAL - Adds granular result types for client queries.
+	/** EXPERIMENTAL - Adds granular result types for queries.
 	Granular here means specifically the required Unreal components for spawning other actors and all data type components.
 	Needs testing thoroughly before making default. May be replaced by component set result types instead. */
 	UPROPERTY(config)
-	bool bEnableClientResultTypes;
+	bool bEnableResultTypes;
 
 	/** Pack RPCs sent during the same frame into a single update. */
 	UPROPERTY(config)
@@ -191,18 +203,6 @@ public:
 	/** The receptionist host to use if no 'receptionistHost' argument is passed to the command line. */
 	UPROPERTY(EditAnywhere, config, Category = "Local Connection")
 	FString DefaultReceptionistHost;
-
-	/** If the Development Authentication Flow is used, the client will try to connect to the cloud rather than local deployment. */
-	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-	bool bUseDevelopmentAuthenticationFlow;
-
-	/** The token created using 'spatial project auth dev-auth-token' */
-	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-	FString DevelopmentAuthenticationToken;
-
-	/** The deployment to connect to when using the Development Authentication Flow. If left empty, it uses the first available one (order not guaranteed when there are multiple items). The deployment needs to be tagged with 'dev_login'. */
-	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-	FString DevelopmentDeploymentToConnect;
 
 	UPROPERTY(EditAnywhere, Config, Category = "Region settings", meta = (ConfigRestartRequired = true, DisplayName = "Region where services are located"))
 	TEnumAsByte<EServicesRegion::Type> ServicesRegion;
@@ -284,4 +284,26 @@ public:
 	bool bAsyncLoadNewClassesOnEntityCheckout;
 
 	FORCEINLINE bool IsRunningInChina() const { return ServicesRegion == EServicesRegion::CN; }
+
+	/** Enable to use the new net cull distance component tagging form of interest */
+	UPROPERTY(EditAnywhere, Config, Category = "Interest")
+	bool bEnableNetCullDistanceInterest;
+
+	/** Enable to use interest frequency with bEnableNetCullDistanceInterest*/
+	UPROPERTY(EditAnywhere, Config, Category = "Interest", meta = (EditCondition = "bEnableNetCullDistanceInterest"))
+	bool bEnableNetCullDistanceFrequency;
+
+	/** Full update frequency ratio of actor's net cull distance */
+	UPROPERTY(EditAnywhere, Config, Category = "Interest", meta = (EditCondition = "bEnableNetCullDistanceFrequency"))
+	float FullFrequencyNetCullDistanceRatio;
+
+	/** QBI pairs for ratio of - net cull distance : update frequency */
+	UPROPERTY(EditAnywhere, Config, Category = "Interest", meta = (EditCondition = "bEnableNetCullDistanceFrequency"))
+	TArray<FDistanceFrequencyPair> InterestRangeFrequencyPairs;
+
+public:
+	// UI Hidden settings passed through from SpatialGDKEditorSettings
+	bool bUseDevelopmentAuthenticationFlow;
+	FString DevelopmentAuthenticationToken;
+	FString DevelopmentDeploymentToConnect;
 };