+++ SpatialGDK/Source/SpatialGDK/Public/Utils/SchemaDatabase.h
@@ -93,6 +93,12 @@ public:
 	TMap<FString, uint32> LevelPathToComponentId;
 
 	UPROPERTY(Category = "SpatialGDK", VisibleAnywhere)
+	TMap<float, uint32> NetCullDistanceToComponentId;
+
+	UPROPERTY(Category = "SpatialGDK", VisibleAnywhere)
+	TSet<uint32> NetCullDistanceComponentIds;
+
+	UPROPERTY(Category = "SpatialGDK", VisibleAnywhere)
 	TMap<uint32, FString> ComponentIdToClassPath;
 
 	// These component ID lists for each data type are stored separately as you cannot have nested maps in a UPROPERTY