+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased-`x.y.z`] - 2020-xx-xx
 
+### Breaking Changes:
+- Simulated Player worker configurations now require a dev auth token and deployment flag instead of a login token and player identity token. See the Example Project for an example of how to set this up.
+
 ### Features:
 - Updated the version of the local API service used by the UnrealGDK.
-- The GDK now uses SpatialOS `14.3.0`.
+- The GDK now uses SpatialOS `14.4.0`.
 - In local deployments of the Example Project you can now launch Simulated Players in one click. Running `LaunchSimPlayerClient.bat` will launch a single Simulated Player client. Running `Launch10SimPlayerClients.bat` will launch 10.
 - Added an AuthorityIntent component to be used in the future for UnrealGDK code to control loadbalancing.
 - Added support for the UE4 Network Profile to measure relative size of RPC and Actor replication data.
@@ -39,9 +42,13 @@ Usage: `DeploymentLauncher createsim <project-name> <assembly-name> <target-depl
 - Entries in the SchemaDatabase are now sorted to improve efficiancy when browsing the asset in the editor. (DW-Sebastien)
 - Load Balancing Strategies and Locking Strategies can be set per-level using SpatialWorldSettings.
 - Batch Spatial Position Updates now defaults to false.
-- Introduced experimental feature flag `bEnableClientResultTypes`, defaulting false. Flip this to true for client Interest queries to only include the set of components required to run. Should give bandwidth savings depending on your game. 
 - Dynamic interest overrides are disabled if the `bEnableClientResultTypes` flag is set to true. 
-
+- Added `bEnableNetCullDistanceInterest` (defaulted false) to enable client interest to be exposed through component tagging. This functionality has closer parity to native unreal client interest.
+- Added `bEnableNetCullDistanceFrequency` (defaulted false) to enable client interest queries to use frequency. This functionality is configured using `InterestRangeFrequencyPairs` and `FullFrequencyNetCullDistanceRatio`.
+- Added support for Android.
+- Introduced experimental feature flag `bEnableResultTypes`, defaulting false. Flip this to true for Interest queries to only include the set of components required to run. Should give bandwidth savings depending on your game. 
+- Moved Dev Auth settings from runtime settings to editor settings.
+- Added the option to use the development authentication flow using the command line.
 
 ## Bug fixes:
 - Fixed a bug that caused the local API service to memory leak.
@@ -71,6 +78,7 @@ Usage: `DeploymentLauncher createsim <project-name> <assembly-name> <target-depl
 - Track properties containing references to replicated actors, in order to resolve them again if the actor they reference moves out and back into relevance.
 - Fix problem where PIE sessions sometimes fail to start due to missing schema for SpatialDebugger blueprint.
 - Fixed an issue where newly created subobjects would have empty state when RepNotify was called for a property pointing to that subobject.
+- Fixed an issue where deleted, initially dormant startup actors would still be present on other workers.
 
 ### External contributors:
 @DW-Sebastien